% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadSearchedTaxa.R
\name{APItaxon}
\alias{APItaxon}
\title{Encode Taxon Name}
\usage{
APItaxon(taxon)
}
\arguments{
\item{taxon}{Single taxonomic name}
}
\value{
Returns an encoded taxon name.
}
\description{
This function will take a taxon name and encode spaces, and underscores to +
for the EOL API.
}
\examples{

	APItaxon("Ursus americanus")

}
