% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/can_use_eol.R
\name{can_use_eol}
\alias{can_use_eol}
\title{Check if the Encyclopedia Of Life website can be used by using a
simple API call}
\usage{
can_use_eol()
}
\value{
TRUE if the Encyclopedia Of Life website can be used, 
FALSE otherwise
}
\description{
Check if the Encyclopedia Of Life website can be used by using a
simple API call
}
\author{
Richel J.C. Bilderbeek
}
