% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataforWeb.R
\name{CollectDataforWeb}
\alias{CollectDataforWeb}
\alias{DataProcessing}
\alias{CNCount}
\alias{DOCount}
\alias{providerCount}
\title{Collect Data from EOL Pages for Website}
\usage{
CollectDataforWeb(MyEOL, do.higher.taxonomy = FALSE)
}
\arguments{
\item{MyEOL}{A single filename for downloaded EOL pages}

\item{do.higher.taxonomy}{TRUE/FALSE on whether to download higher level
taxonomy}

\item{res}{XML object}
}
\value{
Appends EOL data to a table.
}
\description{
These functions will read and scrape content off the downloaded EOL pages.
This is for building the data table on eoldata.org.
}
\examples{

	data(MyEOLs)
	CollectDataforWeb(MyEOLs[1])

}
\author{
Barb Banbury, Brian O'Meara
}
