% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeHierarchyTree.R
\name{CombineHierarchyInfo}
\alias{CombineHierarchyInfo}
\title{Combine Hierarchy Info}
\usage{
CombineHierarchyInfo(MyHiers)
}
\arguments{
\item{MyHiers}{A vector of hier pages to combine}
}
\value{
Returns a combined data frame.
}
\description{
This function combines different hierarchy dataframes into one. Mostly an
internal function for tree and edge label building.
}
\examples{

	data(MyHiers)
	CombineHierarchyInfo(MyHiers)

}
