% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_search_url.R
\name{create_search_url}
\alias{create_search_url}
\title{Create a URL to search the Encyclopedia of Life}
\usage{
create_search_url(species_name, use_exact_name = TRUE)
}
\description{
Create a URL to search the Encyclopedia of Life
}
