% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataforWeb.R
\name{DataProcessing}
\alias{DataProcessing}
\title{Do something}
\usage{
DataProcessing(res, do.higher.taxonomy)
}
\arguments{
\item{res}{something}

\item{do.higher.taxonomy}{somethinf}
}
\description{
Do something
}
\author{
Barb Banbury, Brian O'Meara
}
