% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_image_urls_from_id.R
\name{download_image_urls_from_id}
\alias{download_image_urls_from_id}
\title{Get the URLs of the Encyclopedia Of Life images
for that species}
\usage{
download_image_urls_from_id(species_id)
}
\value{
a URL if there is a URL, an empty string otherwise
}
\description{
Get the URLs of the Encyclopedia Of Life images
for that species
}
\examples{
species_id <- download_search_results("Homo sapiens")$id[1]
urls <- download_image_urls_from_id(species_id)
}
