% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_image_urls_from_name.R
\name{download_image_urls_from_name}
\alias{download_image_urls_from_name}
\title{Get the Encyclopedia Of Life image URL from a species name}
\usage{
download_image_urls_from_name(species_name, use_exact_name = TRUE,
  verbose = FALSE)
}
\arguments{
\item{species_name}{name of a species}

\item{use_exact_name}{do use the exact name}

\item{verbose}{give more output}
}
\value{
a data frame with species name, ID and URLs
}
\description{
Get the Encyclopedia Of Life image URL from a species name
}
