% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_search_results.R
\name{download_search_results}
\alias{download_search_results}
\title{Get the Encyclopedia Of Life page for a certain species}
\usage{
download_search_results(species_name, use_exact_name = TRUE,
  verbose = FALSE)
}
\arguments{
\item{species_name}{name of the species}

\item{use_exact_name}{only return an EOL page if the species name
matches exactly}

\item{verbose}{be verbose}
}
\value{
a data frame with columns \code{id}, \code{title}, \code{link}
and \code{description}
}
\description{
Get the Encyclopedia Of Life page for a certain species
}
\examples{
library(testthat)

df <- download_search_results(species_name = "Homo sapiens")

expect_equal(class(df), "data.frame")
expect_true("id" \%in\% names(df))
expect_true("title" \%in\% names(df))
expect_true("link" \%in\% names(df))
expect_true("content" \%in\% names(df))
}
