% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eol_image_urls.R
\name{get_image_urls_from_id}
\alias{get_image_urls_from_id}
\title{Get the URLs of the Encyclopedia Of Life images
for that species}
\usage{
get_image_urls_from_id(species_id)
}
\value{
a URL if there is a URL, an empty string otherwise
}
\description{
Get the URLs of the Encyclopedia Of Life images
for that species
}
\examples{
species_id <- get_eol_page("Homo sapiens")$id[1]
urls <- get_image_urls_from_id(species_id)
}
