% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataforWeb.R, R/PageProcessing.R
\name{PageProcessing}
\alias{PageProcessing}
\alias{RemoveNAFiles}
\title{Process XML Data into tree format}
\usage{
PageProcessing(MyEOL, ...)

PageProcessing(MyEOL, ...)
}
\arguments{
\item{MyEOL}{A filename or R object for downloaded EOL pages}

\item{...}{further arguments to be passed to \code{PageProcessing}}

\item{MyFiles}{A vector of filenames or R objects for EOL or Hier pages}

\item{MyEOL}{A filename or R object for downloaded EOL pages}

\item{MyFiles}{A vector of filenames or R objects for EOL or Hier pages}
}
\value{
an XML tree as an R object.

\code{PageProcessing} returns XML tree as an R object.
\code{RemoveNAFiles} is an internal function that will take a vector of
filenames and check to make sure they all have data associated (sometimes
requested pages will be empty and contain no information).  If they are
empty, they are cleaved from the analysis.
}
\description{
This function will read in the XML data and parse it into a tree structure for R to read.

This function will read in the XML data and parse it into a tree structure
for R to read.
}
\note{
\code{RemoveNAFiles} is an internal function
that will take a vector of filenames
and check to make sure they all have
data associated (sometimes requested pages
will be empty and contain no information).
If they are empty, they are cleaved from the analysis.
}
\examples{
# Reads in a file
PageProcessing("eol1.xml")

# or reads in an R object
data(MyEOLs)
PageProcessing(MyEOLs[1])#'

	#Reads in a file
	#PageProcessing("eol1.xml") 
	
	#or reads in an R object
	data(MyEOLs)
	PageProcessing(MyEOLs[1])


}
