% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ping_eol.R
\name{ping_eol}
\alias{ping_eol}
\title{Ping the Encyclopedia Of Life}
\usage{
ping_eol(key = NULL)
}
\arguments{
\item{key}{Encyclopedia Of Life identification key}
}
\value{
the word \code{Success} (taken from the Encyclopedia Of Life's
XML response), will \link{stop} otherwise
}
\description{
A 'ping' is a simple call to indicate if an API is in working
order.
}
\examples{
library(testthat)

result <- ping_eol()
expect_equal(ping_eol(), "Success")
}
\seealso{
use \link{can_use_eol} to get a boolean to indicate the
Encylopedia Of Life is in working order
}
