% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CollectDataforWeb.R
\name{providerCount}
\alias{providerCount}
\title{Provider numbers}
\usage{
providerCount(res)
}
\arguments{
\item{res}{something}
}
\description{
Provider numbers
}
\author{
Barb Banbury, Brian O'Meara
}
