% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reol-package.R
\docType{package}
\name{reol}
\alias{reol}
\alias{reol-package}
\title{reol}
\description{
R interface to the Encyclopedia of Life
}
\details{
\tabular{ll}{ Package: \tab Reol\cr Type: \tab Package\cr Version: \tab
1.21\cr Date: \tab 2013-9-9\cr License: \tab GPL (>= 2)\cr}
}
\author{
Barb Banbury, Brian O'Meara, Richel J.C. Bilderbeek
}
