% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_url_friendly_format.R
\name{to_url_friendly_format}
\alias{to_url_friendly_format}
\title{Convert a human written name to something that can fit in a URL}
\usage{
to_url_friendly_format(species_name)
}
\arguments{
\item{species_name}{name of the species}
}
\description{
Convert a human written name to something that can fit in a URL
}
\author{
Richel J.C. Bilderbeek
}
