% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ping_url.R
\name{create_ping_url}
\alias{create_ping_url}
\title{Create the URL to ping the Encyclopedia of Life}
\usage{
create_ping_url(key = NULL)
}
\description{
Create the URL to ping the Encyclopedia of Life
}
\author{
Richel J.C. Bilderbeek
}
