% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_searches_results.R
\name{download_searches_results}
\alias{download_searches_results}
\title{Get the Encyclopedia Of Life pages for multiple taxa}
\usage{
download_searches_results(taxon_names)
}
\value{
a list with as much elements as taxon names. Each
element is a data frame
}
\description{
Get the Encyclopedia Of Life pages for multiple taxa
}
\author{
Richel J.C. Bilderbeek
}
