% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eol_pages.R
\name{get_eol_pages}
\alias{get_eol_pages}
\title{Get the Encyclopedia Of Life pages for multiple taxa}
\usage{
get_eol_pages(taxon_names)
}
\value{
a list with as much elements as taxon names. Each
element is a data frame
}
\description{
Get the Encyclopedia Of Life pages for multiple taxa
}
\author{
Richel J.C. Bilderbeek
}
