% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_eol_url.R
\name{get_eol_url}
\alias{get_eol_url}
\title{Get the Encyclopedia of Life URL for requesting a species}
\usage{
get_eol_url(species_name, use_exact_name = TRUE)
}
\arguments{
\item{species_name}{name of a species}

\item{use_exact_name}{do use the exact name}
}
\description{
Get the Encyclopedia of Life URL for requesting a species
}
